Change to use ReadSB as default receiver
Since most people will be using ReadSB as they will be coming from the guide, without this change they need to manually edit the config file to get any data. This will run every time the container starts, but will only match once. Most users will not use 127.0.0.1 since it does not point to the host system and is not useful inside a container.